+++ client/src/pages/studyList/StudyListView.js
@@ -30,54 +30,59 @@ const StudyListView = props => {
 
   const user = useSelector(state => state.user.value);
   const filters = useSelector(state => state.filters.value);
-  const [state, setState] = useState({
-    isLoaded: false,
-    isError: false,
-    data: {}
-  });
+  const [studyData, setStudyData] = useState(null);
+  const [error, setError] = useState(null);
+
+  const indexStudies = (studies) => {
+    const data = {};
+    data.cf = crossfilter(studies);
+    data.dimensions = {};
+    data.dimensions.allData = data.cf.dimension(d => d);
+    data.dimensions[filter.PROGRAM] = data.cf.dimension(d => d.program.id);
+    data.dimensions[filter.LEGACY] = data.cf.dimension(d => d.legacy);
+    data.dimensions[filter.EXTERNAL] = data.cf.dimension(
+        d => !!d.collaborator);
+    data.dimensions[filter.MY_STUDY] = data.cf.dimension(d => {
+      if (!!user) {
+        if (d.owner.id === user.id) return true;
+        else {
+          for (const u of d.users) {
+            if (u.id === user.id) return true;
+          }
+        }
+      }
+      return false;
+    });
+    data.dimensions[filter.STATUS] = data.cf.dimension(d => d.status);
+    data.dimensions[filter.ACTIVE] = data.cf.dimension(d => d.active);
+    setStudyData(data);
+  }
+
+  const applyFilters = (filters) => {
+    for (let key of Object.keys(studyData.dimensions)) {
+      studyData.dimensions[key].filterAll();
+      if (filters.hasOwnProperty(key) && filters[key] != null) {
+        console.debug("Applying filter", key, filters[key]);
+        if (Array.isArray(filters[key])) {
+          studyData.dimensions[key].filter(
+              d => filters[key].indexOf(d) > -1);
+        } else {
+          studyData.dimensions[key].filter(filters[key]);
+        }
+      }
+    }
+  }
 
   useEffect(() => {
 
     axios.get("/api/internal/study")
     .then(response => {
-
       console.debug("Studies", response.data);
-
-      const data = {};
-      data.cf = crossfilter(response.data);
-      data.dimensions = {};
-      data.dimensions.allData = data.cf.dimension(d => d);
-      data.dimensions[filter.PROGRAM] = data.cf.dimension(d => d.program.id);
-      data.dimensions[filter.LEGACY] = data.cf.dimension(d => d.legacy);
-      data.dimensions[filter.EXTERNAL] = data.cf.dimension(
-          d => !!d.collaborator);
-      data.dimensions[filter.MY_STUDY] = data.cf.dimension(d => {
-        if (!!user) {
-          if (d.owner.id === user.id) return true;
-          // else {
-          //   for (const u of d.users) {
-          //     if (u.id === user.id) return true;
-          //   }
-          // }
-        }
-        return false;
-      });
-      data.dimensions[filter.STATUS] = data.cf.dimension(d => d.status);
-
-      setState(prevState => ({
-        ...prevState,
-        data: data,
-        isLoaded: true
-      }));
-
+      indexStudies(response.data);
     })
     .catch(error => {
       console.error(error);
-      setState(prevState => ({
-        ...prevState,
-        isError: true,
-        error: error
-      }));
+      setError(error);
     });
   }, [user]);
 
@@ -86,34 +91,23 @@ const StudyListView = props => {
 
   try {
 
-    if (state.isError) {
+    if (error) {
 
       content = <ErrorMessage/>;
 
-    } else if (state.isLoaded) {
+    } else if (studyData) {
 
       // Apply filters
       console.debug("Active filters", filters);
+      applyFilters(filters);
 
-      for (let key of Object.keys(state.data.dimensions)) {
-        state.data.dimensions[key].filterAll();
-        if (filters.hasOwnProperty(key) && filters[key] != null) {
-          console.debug("Applying filter", key, filters[key]);
-          if (Array.isArray(filters[key])) {
-            state.data.dimensions[key].filter(
-                d => filters[key].indexOf(d) > -1);
-          } else {
-            state.data.dimensions[key].filter(filters[key]);
-          }
-        }
-      }
-
-      content =
+      content = (
           <StudyList
-              studies={state.data.dimensions.allData.top(Infinity)}
+              studies={studyData.dimensions.allData.top(Infinity)}
               filters={filters}
               user={user}
-          />;
+          />
+      );
 
     }